設定を別ページで行うように修正
