+++ lib/view/setting_view.dart
@@ -1,6 +1,14 @@
 import 'package:flutter/material.dart';
 
-import '../component/setting/theme_setting.dart';
+import 'setting/theme_setting.dart';
+
+final List<Map<String, Object>> settingList = [
+  {
+    'title': '外観モード',
+    'icon': Icons.contrast,
+    'view': const ThemeSetting(),
+  },
+];
 
 class SettingPage extends StatelessWidget {
   const SettingPage({super.key});
@@ -8,10 +16,25 @@ class SettingPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: Column(
-        children: <Widget>[
-          const ThemeSetting(),
-        ],
+      body: ListView.builder(
+        itemCount: settingList.length,
+        itemBuilder: (context, index) {
+          Map<String, Object> settingItem = settingList[index];
+          return ListTile(
+            leading: Icon(settingItem['icon'] as IconData),
+            title: Text(settingItem['title'] as String),
+            trailing: const Icon(Icons.navigate_next),
+            onTap: () {
+              Navigator.of(context).push<void>(
+                MaterialPageRoute(
+                  builder: (context) {
+                    return settingItem['view'] as Widget;
+                  },
+                ),
+              );
+            },
+          );
+        },
       ),
     );
   }
